feat: 메인페이지 유저 프로필 컴포넌트 분리 및 정보 추가
