+++ app/page.tsx
@@ -2,12 +2,10 @@
 
 import { useState } from 'react';
 
-import { Image } from '@/components/atoms';
-import { Calendar } from '@/features/main/calendar';
+import { UserCalendarProfile } from '@/features/main/calendar';
 import { MainTab } from '@/features/main/main-tab';
 import { TabItemInfo } from '@/features/main/main-tab/main-tab';
 import { TimeLine } from '@/features/main/time-line';
-import { css } from '@/styled-system/css';
 import { flex } from '@/styled-system/patterns';
 
 const SELECT_CALENDAR_VIEW = 0;
@@ -15,7 +13,6 @@ const SELECT_TIMELINE_VIEW = 1;
 
 export default function Home() {
   const [selected, setSelected] = useState<number>(SELECT_CALENDAR_VIEW);
-
   const tabInfos: Array<TabItemInfo> = [
     {
       text: '캘린더',
@@ -36,29 +33,7 @@ export default function Home() {
         <MainTab tabInfos={tabInfos} />
       </header>
       <section className={contentStyles}>
-        {isSelectedCalendarView ? (
-          <>
-            <div className={profileContainerStyles}>
-              <Image
-                className={characterImageStyles}
-                width={70}
-                height={75}
-                src="/images/swimie-character.svg"
-                alt="swimie character"
-                priority
-              />
-              <div className={userInfoStyles}>
-                <p className={nicknameStyles}>수린이님,</p>
-                <p className={descriptionStyles}>
-                  이번달 수영을 7번 다녀왔어요!
-                </p>
-              </div>
-            </div>
-            <Calendar />
-          </>
-        ) : (
-          <TimeLine />
-        )}
+        {isSelectedCalendarView ? <UserCalendarProfile /> : <TimeLine />}
       </section>
     </main>
   );
@@ -77,32 +52,4 @@ const contentStyles = flex({
   padding: '16px 0',
   gap: '16px',
   direction: 'column',
-});
-
-const profileContainerStyles = flex({
-  padding: '14px 11px',
-  gap: '4px',
-  alignItems: 'center',
-  borderRadius: '6px',
-  backgroundColor: 'primary.swim.총거리.default',
-});
-
-const characterImageStyles = css({ margin: '0 10px' });
-
-const userInfoStyles = flex({
-  height: 'full',
-  direction: 'column',
-  justifyContent: 'center',
-});
-
-const nicknameStyles = css({
-  textStyle: 'heading6',
-  fontWeight: 'bold',
-  color: 'background.white',
-});
-
-const descriptionStyles = css({
-  textStyle: 'label1.normal',
-  fontWeight: 'medium',
-  color: 'line.solid.neutral',
 });
