+++ features/main/calendar/index.ts
@@ -1,2 +1,3 @@
 export * from './molecules';
+export * from './organisms';
 export * from './types';
