+++ features/main/calendar/organisms/index.ts
@@ -1,0 +1,1 @@
+export * from './user-calendar-profile';
